Release v1.8.1 - bugfixes & hardening
- Authentic SID audio export (libsidplayfp -> WAV/FLAC/OGG/MP3) when the C64
  Emulation engine is active; no SoundFont needed, whole-tune render.
- C64 fixes: engine-switch crash during playback, SoundFont-mode export timbres,
  Emulation gated on a loaded .sid, stale C64/FFXIV SoundFont now falls back to GM/GS.
- Code-review hardening: GPX BCFZ OOB read, three playback-thread data races
  (bard-note bookkeeping, FFXIV equalizer table, SID renderer mute), recording
  orphan-note leak, MusicXML duration overflow, AutoUpdater brick-on-failed-extract,
  FFXIV EQ button enable, GP repeat-alternative UB shift, tempo-map null guard.
- Smaller Windows download: drop the unused Qt FFmpeg media backend (~15 MB).
- Disabled checkboxes / radio buttons now show muted text in all themes.
- download.html version/size now fetched live from the GitHub API.

+++ src/converter/MusicXml/MusicXmlImporter.cpp
@@ -182,8 +182,12 @@ private:
 
         const int globalPPQ = _score->ticksPerQuarter;
         auto toGlobal = [&](int xmlDur) {
-            // Scale local divisions to global PPQ.
-            return divisions > 0 ? (xmlDur * globalPPQ) / divisions : xmlDur;
+            // Scale local divisions to global PPQ. 64-bit intermediate so a large
+            // <duration>/<backup>/<forward> can't overflow 32-bit int
+            // (xmlDur * globalPPQ wrapped before, BUG-CORE-008).
+            return divisions > 0
+                ? static_cast<int>((static_cast<qint64>(xmlDur) * globalPPQ) / divisions)
+                : xmlDur;
         };
 
         while (!xr.atEnd()) {
